chore: factorize header importation
import-headers and bootstrap both use the same peerless tactic
to import headers from source `data` directory.
